requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
passlib[bcrypt]
pyjwt
python-multipart
pydantic
pybreaker
tenacity
email-validator